refactor: new SizeRotator
